move previews to separate file
